move input props interfaces to new file
